feat: move addbottochannel to bot.tsx
